Fix two problems with new feedback()
1. Add an else clause for when the yellow letter shouldn't get added, so
that a dash gets added instead

2. Initalize `response` to only contain spaces and pipe instead of
inititalizing it with the guess on the left. Including the guess made
it hard to check how many times the current letter appeared in the
response so far - we only want to count the letters of the feedback
that's being constructed, and NOT check the letters of the guess.
So instead, the guess and the feedback get concatenated in the return.